Extract the CLI entry points into cgt_calc/cli.py
First step of splitting main.py into phase modules. Moves calculate_cgt,
main and init to a new cgt_calc/cli.py, points the cgt-calc console script
at cgt_calc.cli:init, and leaves re-exports and the __main__ guard in
cgt_calc.main so existing imports and python -m cgt_calc.main keep
working. Pure move, no behavior change; one monkeypatch target in
test_calc.py follows calculate_cgt to its new module.
